+++ client/src/components/Calendars.js
@@ -1,0 +1,14 @@
+import React, { useState } from "react";
+import { Calendar, theme } from "antd";
+
+export const DoctorCalendar = () => {
+  return (
+    <>
+      <Calendar fullscreen={false} />
+    </>
+  );
+};
+
+export const PatientCalendar = () => {
+  return <>This is Patient's Calendar</>;
+};
